backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pytest==8.0.0
pytest-asyncio==0.23.5
supabase-py==2.3.4
httpx[http2]==0.26.0
tenacity==8.2.3 
redis==5.0.0
python-dotenv==1.0.1
